DOMA-1437 move canReadSensetive data to accessSchema
